requirements.txt
asgiref==3.9.1
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.5
gunicorn==23.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.2
packaging==25.0
pillow==11.3.0
proglog==0.1.12
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tqdm==4.67.1
